+++ opentmf-api-clients-reactive/src/main/java/org/opentmf/api/client/reactive/impl/GenericReactiveTmfClientImpl.java
@@ -2,6 +2,7 @@ package org.opentmf.api.client.reactive.impl;
 
 import org.opentmf.api.client.common.config.TmfApiClientsConfig.EndpointConfig;
 import org.opentmf.api.client.common.config.TmfApiClientsConfig.ServerConfig;
+import org.opentmf.api.client.common.model.SubResourcePath;
 import org.opentmf.api.client.reactive.api.GenericReactiveTmfClient;
 import org.opentmf.client.common.model.ClientProperties;
 import org.opentmf.client.reactive.service.api.TokenService;
@@ -21,5 +22,16 @@ public class GenericReactiveTmfClientImpl
       TokenService tokenService,
       ClientProperties clientProperties) {
     super(endpointConfig, serverConfig, webClient, tokenService, clientProperties, Object.class);
+  }
+
+  public GenericReactiveTmfClientImpl(
+      EndpointConfig endpointConfig,
+      ServerConfig serverConfig,
+      WebClient webClient,
+      TokenService tokenService,
+      ClientProperties clientProperties,
+      SubResourcePath subPath) {
+    super(endpointConfig, serverConfig, webClient, tokenService, clientProperties, Object.class,
+        subPath);
   }
 }