+++ opentmf-api-clients-rest/src/main/java/org/opentmf/api/client/rest/impl/GenericTmfClientImpl.java
@@ -2,6 +2,7 @@ package org.opentmf.api.client.rest.impl;
 
 import org.opentmf.api.client.common.config.TmfApiClientsConfig.EndpointConfig;
 import org.opentmf.api.client.common.config.TmfApiClientsConfig.ServerConfig;
+import org.opentmf.api.client.common.model.SubResourcePath;
 import org.opentmf.api.client.rest.api.RestGenericTmfClient;
 import org.opentmf.client.common.model.ClientProperties;
 import org.opentmf.client.rest.service.api.SyncTokenService;
@@ -21,5 +22,16 @@ public class GenericTmfClientImpl
       SyncTokenService tokenService,
       ClientProperties clientProperties) {
     super(endpointConfig, serverConfig, restClient, tokenService, clientProperties, Object.class);
+  }
+
+  public GenericTmfClientImpl(
+      EndpointConfig endpointConfig,
+      ServerConfig serverConfig,
+      RestClient restClient,
+      SyncTokenService tokenService,
+      ClientProperties clientProperties,
+      SubResourcePath subPath) {
+    super(endpointConfig, serverConfig, restClient, tokenService, clientProperties, Object.class,
+        subPath);
   }
 }